requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
duckdb==1.1.3
et_xmlfile==2.0.0
Faker==19.13.0
fonttools==4.55.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watson_machine_learning==1.0.367
ibm_watsonx_ai==1.1.23
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-ibm==0.3.3
langchain-text-splitters==0.3.2
langgraph==0.2.50
langgraph-checkpoint==2.0.4
langgraph-sdk==0.1.36
langsmith==0.1.143
lomond==0.3.3
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
numpy==1.26.4
openai==1.54.4
openpyxl==3.1.5
orjson==3.10.11
packaging==24.2
pandas==1.5.3
pandasai==2.3.0
pillow==10.4.0
propcache==0.2.0
psycopg2==2.9.10
pyarrow==18.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.2
requests-toolbelt==1.0.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlglot==25.30.0
sqlglotrs==0.2.13
sqlparse==0.5.2
tabulate==0.9.0
tenacity==9.0.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
zipp==3.21.0
